use stringReplacement instead of extras
